switch to client()
For producing with zip, we'll add an opt in the run command.

Doing so requires not having the client initialized yet so we can stuff
an opt in as appropriate.

So, we'll switch all client usages to load on the first use of the
client.

Also fixes compile bugs from the prior commit in producing; didn't
build.