beginnings of a refactor; consume moved
I didn't plan on this refactor _now_, but consuming is getting unweildy
and I want to packagize things. In the process of moving the consumer I
decided to move the regex and topic stuff directly into kgo since it is
hideous here; that is yet uncommitted but I want to finish the refactor
to test it. So, that begins now.