format parsing in: allow plain ascii number parsing
This can be a bit of a footgun but can also be pretty convenient.

e.g. the following is now valid:

  echo "2ov 2 1k1v1k1v" | kcl produce foo -f '%V%v %H %h{%K%k%V%v}\n'

This will parse a two byte value, two headers, and each header has a one
byte key (k) and one byte value (v).
I was using this to check that it's sound to duplicate headers, and
indeed it is.

The footgun aspect arises from if you do something like

  echo "2ov21k1v1k1v" | kcl produce foo -f '%V%v%H%h{%K%k%V%v}\n'

i.e. remove the spaces around the header count. This will be parsed as
21 headers, and then will fail. Similarly, if the string you are parsing
begins with a number, that number will be parsed as the size.