probe-version: fix potential problem for identifying old kafkas
We cannot reuse the same client if the kafka is pre 0.10.0, since the
client will still try issuing ApiVersions first thing when it connects
to a broker.

Instead, we recreate the client with a new lower MaxVersions.

This makes probe-versions incompatible with --as-version, since the
remade client will override that flag. This is fine, since if we are
probing for versions, it doesn't make sense to also have a set max
version.