update for latest kgo updates
This converts relevant requests to use RequestSharded, so that now the
output contains which broker the data came from where relevant.

This avoids using RequestSharded for alter configs / describe configs,
because that'd be more plumbing than I care for right now, but more
importantly, we only ever have one description / alter per request, so
the request only ever goes to one broker. With -tb, the user is
specifying the broker anyway.

kcl misc list-offsets may be faster now because we no longer wait to
issue the end offsets request until after the start finishes; instead we
list both concurrently.

kcl myconfig ls was expanded such that the symlink and currently used
config are both annotated now.

The client has been updated such that 2.6.0 and 2.7.0 are options for
the max version stuff.