fix(myconfig): keep tls when the struct is empty but the tls is required
In servers like Confluent Kafka TLS and SASL is required, but is not required
configure a custom client/ca certificate, so is only necessary keep the `[tls]`
on toml file.

The omitzero will keep the key `[tls]` if the CfgTLS is not nil but all values
of the struct is empty.